adds TodoList component
