+++ src/App.jsx
@@ -1,12 +1,6 @@
 // import { useState } from 'react'
 import './App.css'
-
-const todoList = [];
-[
-  'Master React', 
-  'Master Node', 
-  'Clear out email'
-].forEach((title, id) => { todoList.push({id, title}) });
+import TodoList from './TodoList.jsx';
 
 function App() {
   // const [count, setCount] = useState(0)
@@ -14,17 +8,7 @@ function App() {
   return (
     <>
       <h1>Todo List</h1>
-      <ul className='todo-list'>
-        {
-          todoList.map(({title, id}) => (
-            <li
-              key={id}
-            >
-              {title}
-            </li>
-          ))
-        }
-      </ul>
+      <TodoList />
     </>
   )
 }
